Implementing voyager screen implementation for ProductsScreen.
