--- shared/src/commonMain/kotlin/me/inassar/features/feature/presentation/ui/FeatureScreen.kt
@@ -1,26 +1,0 @@
-package me.inassar.features.feature.presentation.ui
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
-import dev.icerock.moko.mvvm.compose.getViewModel
-import dev.icerock.moko.mvvm.compose.viewModelFactory
-import me.inassar.features.feature.presentation.manipulator.FeatureEvents
-import me.inassar.features.feature.presentation.manipulator.FeatureViewModel
-
-/**
- * Created by Ahmed Nassar on 5/27/23.
- */
-
-@Composable
-fun ProductsScreen() {
-    val viewModel: FeatureViewModel = getViewModel(key = Unit,
-        factory = viewModelFactory { FeatureViewModel() })
-    val productsState = viewModel.productsState.collectAsState().value
-
-    LaunchedEffect(Unit) {
-        viewModel.onEvent(FeatureEvents.GetProducts)
-    }
-
-    ProductsList(productsState)
-}
